Merge from JMRI Master

+++ scripts/diskimage-sign.sh
@@ -35,12 +35,15 @@ AC_USER=$5
 AC_PASSWORD=$6
 KEYCHAIN_FILE=$7
 
+JAR=/usr/bin/jar
+
 # -----------------------------------------
 function trapExitHandler {
     trap - 1 2 3 15
     umount "$tmpimage2" && echo "Unmounted tmpimage2"
     umount "$tmpindir" && echo "Unmounted tempindir"
-    umount "$INPUTIMAGEFILE" && echo "Unmounted input image"
+    rm -rf "$INPUTIMAGEFILE" && echo "Deleted input image file"
+    umount "$tmpoutdir" && echo "Unmounted tmpoutdir image"    
     rm -rf "$tmpimage1" "$tmpimage2" "$tmpoutdir" "$tmpindir" >/dev/null 2>&1
 }
 # -----------------------------------------
@@ -77,11 +80,15 @@ function signJarMember {
   local jar=$1
   local file=$2
 
-  jar xf $jar $file
-  signFile $file
-  jar uvf $jar $file
-  rm $file
-
+  if [ -f "$jar" ]
+  then
+    export JAVA_HOME=`/usr/libexec/java_home -v 1.8`
+    $JAR xf $jar $file
+    signFile $file
+    $JAR uvf $jar $file
+    rm $file
+  fi
+  
   return 0
 }
 # -----------------------------------------
@@ -90,9 +97,13 @@ function signJarMember {
 function signFile {
   local file=$1
   echo sign $file
-  xattr -lr $file
-  xattr -cr $file
-  codesign -v -s "$CERTIFICATE" --force --keychain "$KEYCHAIN_FILE" --deep $file
+  
+  if [ -e "$file" ]
+  then
+    xattr -lr $file
+    xattr -cr $file
+    sudo -u jake codesign -v -s "$CERTIFICATE" --force --keychain "$KEYCHAIN_FILE" --deep $file
+  fi
   return 0
 }
 # -----------------------------------------
@@ -104,11 +115,23 @@ then
   exit 1
 fi
 
+# switch to a directory with known-good permissions
+if ! workdir=`mktemp -d -t workdir`
+then
+  echo "Cannot create temporary working directory"
+  exit 1
+fi
+cd $workdir
+echo 'workdir' $workdir
+
 if [ -x /usr/bin/hdiutil ]
 then
   # if a Linux box were to have hdiutil, I think that would be the preferable route to follow
   # although it's pretty unlikely
   SYSTEM=MACOSX
+  
+  export JAVA_HOME=`/usr/libexec/java_home -v 1.8`
+  java -version
 else
   SYSTEM=LINUX
 fi
@@ -118,21 +141,28 @@ then
   echo "Cannot create output temporary directory"
   exit 1
 fi
+echo "tmpoutdir" $tmpoutdir
+
 if ! tmpindir=`mktemp -d -t JMRI.input`
 then
   echo "Cannot create input temporary directory"
   exit 1
 fi
+echo 'tmpindir' $tmpindir
+ 
 if ! tmpimage1=`mktemp -t JMRI.tmp.image.1`
 then
   echo "Cannot create temp image 1"
-  exit 1
+  exit
 fi
+echo 'tempimage1' $tmpimage1
+
 if ! tmpimage2=`mktemp -t JMRI.tmp.image.2`
 then
   echo "Cannot create temp image 1"
   exit 1
 fi
+echo 'tmpimage2' $tmpimage2
 
 # handle cleanup on exit
 trap trapExitHandler  0
@@ -146,6 +176,7 @@ rm -f "$tmpimage1" "$tmpimage2"
 # mount input image
 sync
 hdiutil attach "$INPUTIMAGEFILE" -mountpoint "$tmpindir" -nobrowse
+echo "INPUTIMAGEFILE" $INPUTIMAGEFILE
 
 # create disk image and mount
 jmrisize=`du -ms "$tmpindir" | awk '{print $1}'`
@@ -176,11 +207,11 @@ fi
 tar -C "$tmpindir" -cf - JMRI | $SUDO tar -C "$tmpoutdir" -xf -
 
 # display debug info for the keychain containing the certification
-security list-keychains
-security -v default-keychain
-security -v login-keychain
-security -v show-keychain-info "$KEYCHAIN_FILE"
-security -v find-certificate -c "$CERTIFICATE"  "$KEYCHAIN_FILE"
+#security list-keychains
+#security -v default-keychain
+#security -v login-keychain
+#security -v show-keychain-info "$KEYCHAIN_FILE"
+#security -v find-certificate -c "$CERTIFICATE"  "$KEYCHAIN_FILE"
 
 # sign the app files in output
 signFile $tmpoutdir/JMRI/PanelPro.app
@@ -196,11 +227,14 @@ signFile $tmpoutdir/JMRI/lib/macosx/libopenal.1.dylib
 signFile $tmpoutdir/JMRI/lib/macosx/libopenal.dylib
 
 # sign libraries inside jar files
-signJarMember $tmpoutdir/JMRI/lib/libusb4java-1.2.0-osx-x86_64.jar org/usb4java/osx-x86_64/libusb4java.dylib
+signJarMember $tmpoutdir/JMRI/lib/libusb4java-1.3.0-darwin-x86-64.jar org/usb4java/darwin-x86-64/libusb4java.dylib
 signJarMember $tmpoutdir/JMRI/lib/bluecove-2.1.1-SNAPSHOT.jar libbluecove.jnilib
+
 signJarMember $tmpoutdir/JMRI/lib/jna-4.4.0.jar com/sun/jna/darwin/libjnidispatch.jnilib
+signJarMember $tmpoutdir/JMRI/lib/jna-5.9.0.jar com/sun/jna/darwin-x86-64/libjnidispatch.jnilib
+signJarMember $tmpoutdir/JMRI/lib/jna-5.9.0.jar com/sun/jna/darwin-aarch64/libjnidispatch.jnilib
+
 signJarMember $tmpoutdir/JMRI/lib/hid4java-0.5.0.jar darwin/libhidapi.dylib
-# signJarMember $tmpoutdir/JMRI/lib/selenium-server-standalone-3.6.0.jar com/sun/jna/darwin/libjnidispatch.jnilib # OMITTED DUE TO TOC ISSUE
 signJarMember $tmpoutdir/JMRI/lib/jython-standalone-2.7.2.jar META-INF/native/osx/libjansi.jnilib
 signJarMember $tmpoutdir/JMRI/lib/jython-standalone-2.7.2.jar jni/Darwin/libjffi-1.2.jnilib
